botTest/requirements.txt
botcity-framework-core>=0.2.9,<1.0
botcity-maestro-sdk>=0.1.5,<1.0
# Uncomment to use botcity-utils
#botcity-utils>=0.1.2
# Uncomment to use botcity-crawler-plugin
#botcity-crawler-plugin
